+++ src/T2t.Barcode.Svg/Code11.cs
@@ -272,7 +272,7 @@ public class Code11BarcodeDraw
     /// <param name="barMinWidth"></param>
     /// <param name="barMaxWidth"></param>
     /// <returns></returns>
-    protected override int GetDefaultInterGlyphSpace(int barMinWidth, int barMaxWidth)
+    protected override float GetDefaultInterGlyphSpace(float barMinWidth, float barMaxWidth)
     {
         return barMinWidth;
     }
